+++ skills/自媒体运营/发布公众号/__init__.py
@@ -1,0 +1,12 @@
+#!/usr/bin/env python3
+"""微信公众号草稿箱发布工具 — 入口与公共接口。"""
+
+from .client import create_draft, upload_image
+from .converter import markdown_to_html, wrap_full_html
+
+__all__ = [
+    "create_draft",
+    "upload_image",
+    "markdown_to_html",
+    "wrap_full_html",
+]
